requirements.txt
tensorflow
pandas
stringdist
scikit-learn
tqdm
bert-serving-client
bert-serving-server
sentencepiece